+++ examples/counter.rs
@@ -2,49 +2,45 @@ use core::time::Duration;
 
 use async_main::{async_main, LocalSpawner};
 use async_std::task::sleep;
-use pasts::{notify, prelude::*, Loop};
-
-// Exit type for App.
-struct Exit;
+use pasts::{notify, prelude::*};
 
 // Shared state between tasks on the thread.
-struct App<'a> {
+struct App {
     counter: usize,
-    one: &'a mut (dyn Notify<Event = ()> + Unpin),
-    two: &'a mut (dyn Notify<Event = ()> + Unpin),
 }
 
-impl App<'_> {
-    fn one(&mut self, (): ()) -> Poll<Exit> {
+impl App {
+    fn one(&mut self, (): ()) -> bool {
         println!("One {}", self.counter);
         self.counter += 1;
 
         if self.counter > 6 {
-            Ready(Exit)
+            false
         } else {
-            Pending
+            true
         }
     }
 
-    fn two(&mut self, (): ()) -> Poll<Exit> {
+    fn two(&mut self, (): ()) -> bool {
         println!("Two {}", self.counter);
         self.counter += 1;
 
-        Pending
+        true
     }
 }
 
 #[async_main]
 async fn main(_spawner: LocalSpawner) {
     let sleep = |seconds| sleep(Duration::from_secs_f64(seconds));
-    let mut app = App {
-        counter: 0,
-        one: &mut notify::future_fn(|| Box::pin(sleep(1.0))),
-        two: &mut notify::future_fn(|| Box::pin(sleep(2.0))),
-    };
+    let ref mut one = notify::future_fn(|| Box::pin(sleep(1.0)));
+    let ref mut two = notify::future_fn(|| Box::pin(sleep(2.0)));
+    let ref mut app = App { counter: 0 };
 
-    Loop::new(&mut app)
-        .on(|s| s.one, App::one)
-        .on(|s| s.two, App::two)
-        .await;
+    while notify::select([
+        &mut one.map(|()| App::one as fn(&mut App, ()) -> bool),
+        &mut two.map(|()| App::two as _),
+    ])
+    .next()
+    .await(app, ())
+    {}
 }